Switch the type of ChildExecutionInfo.parentClosePolicy from int32 to an enum (#401)

+++ common/persistence/persistenceInterface.go
@@ -1012,7 +1012,7 @@ func (v *InternalChildExecutionInfo) ToProto() *persistenceblobs.ChildExecutionI
 		CreateRequestId:        v.CreateRequestID,
 		Namespace:              v.Namespace,
 		WorkflowTypeName:       v.WorkflowTypeName,
-		ParentClosePolicy:      int32(v.ParentClosePolicy),
+		ParentClosePolicy:      v.ParentClosePolicy,
 	}
 	return info
 }
